+++ footer.php
@@ -16,7 +16,7 @@
 		<?php endif; ?>
 
 		<?php if ( has_nav_menu( 'footer' ) ) : ?>
-			<nav class="footer-navigation" aria-label="<?php esc_attr_e( 'Footer Navigation', 'klaro' ); ?>">
+			<nav class="footer-navigation" aria-label="<?php esc_attr_e( 'Footer', 'klaro' ); ?>">
 				<?php
 				wp_nav_menu(
 					array(